+++ src/ai/regression.h
@@ -8,11 +8,11 @@ typedef struct Regressor
     // 重み（EDGEペア統合分のサイズを減らす)
     //float weights[NB_FEAT_COMB];
     // weight[player][feat_index][pattern_pow3-shape]
-    float *weight[2][FEAT_NUM];
+    double *weight[2][FEAT_NUM];
 #ifdef LEARN_MODE
     uint32_t *nbAppears[FEAT_NUM];
-    float *del[FEAT_NUM];
-    float beta;
+    double *del[FEAT_NUM];
+    double beta;
 #endif
 } Regressor;
 
@@ -20,7 +20,7 @@ void InitRegr(Regressor regr[NB_PHASE]);
 void DelRegr(Regressor regr[NB_PHASE]);
 void RegrClearWeight(Regressor regr[NB_PHASE]);
 void RegrApplyWeightToOpp(Regressor *regr);
-float RegrPred(Regressor *regr, const uint16_t features[], uint8 player);
+double RegrPred(Regressor *regr, const uint16_t features[], uint8 player);
 
 void RegrSave(Regressor regr[NB_PHASE], const char *file);
 void RegrLoad(Regressor regr[NB_PHASE], const char *file);
